# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для управления рекламодателями и их рекламными кампаниями.",
        "title": "Eshke Reklama API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/ad_campaigns": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Кампании рекламодателя из сессии",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ad_campaigns"
                ],
                "summary": "Список рекламных кампаний",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ListAdCampaignsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Создаёт кампанию; рекламодатель определяется по сессии",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ad_campaigns"
                ],
                "summary": "Создание рекламной кампании",
                "parameters": [
                    {
                        "description": "Параметры кампании",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAdCampaignRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAdCampaignResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    }
                }
            }
        },
...
        "/api/ad_campaigns/{ad_campaign_id}/ad_groups/{ad_group_id}/ads/{ad_id}/stats": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает показатели объявления, динамику по дням и разбивку по площадкам",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Статистика объявления",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID кампании",
                        "name": "ad_campaign_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID группы объявлений",
                        "name": "ad_group_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID объявления",
                        "name": "ad_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата начала периода в формате YYYY-MM-DD",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата конца периода в формате YYYY-MM-DD",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.AdStatsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    }
                }
            }
        },
        "/api/ad_campaigns/{ad_campaign_id}/ad_groups/{ad_group_id}/stats": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает показатели группы, динамику по дням и разбивки по объявлениям/площадкам",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Статистика группы объявлений",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID кампании",
                        "name": "ad_campaign_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID группы объявлений",
                        "name": "ad_group_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата начала периода в формате YYYY-MM-DD",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата конца периода в формате YYYY-MM-DD",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GroupStatsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    }
                }
            }
        },
...
        "/api/ad_campaigns/{ad_campaign_id}/stats": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Возвращает показы, клики, CTR, расходы, CPC, вознаграждение партнерам, выручку платформы, динамику по дням и разбивки по группам/площадкам",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Статистика кампании",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID кампании",
                        "name": "ad_campaign_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата начала периода в формате YYYY-MM-DD",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата конца периода в формате YYYY-MM-DD",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CampaignStatsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/httpx.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AdCampaignResponse": {
            "type": "object",
            "properties": {
                "cpm_price": {
                    "type": "integer"
                },
                "daily_budget": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "main_action": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "dto.AdStatsResponse": {
            "type": "object",
            "properties": {
                "period": {
                    "$ref": "#/definitions/dto.StatsPeriod"
                },
                "placements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StatsEntityRow"
                    }
                },
                "previous_totals": {
                    "$ref": "#/definitions/dto.StatsMetric"
                },
                "timeline": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StatsPoint"
                    }
                },
                "totals": {
                    "$ref": "#/definitions/dto.StatsMetric"
                },
                "unavailable_metrics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StatsUnavailableMetric"
                    }
                }
            }
        },
...
        "dto.CampaignStatsResponse": {
            "type": "object",
            "properties": {
                "groups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StatsEntityRow"
                    }
                },
                "period": {
                    "$ref": "#/definitions/dto.StatsPeriod"
                },
                "placements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StatsEntityRow"
                    }
                },
                "previous_totals": {
                    "$ref": "#/definitions/dto.StatsMetric"
                },
                "timeline": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StatsPoint"
                    }
                },
                "totals": {
                    "$ref": "#/definitions/dto.StatsMetric"
                },
                "unavailable_metrics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StatsUnavailableMetric"
                    }
                }
            }
        },
        "dto.CreateAdCampaignRequest": {
            "type": "object",
            "required": [
                "cpm_price",
                "daily_budget",
                "main_action",
                "name"
            ],
            "properties": {
                "cpm_price": {
                    "type": "integer"
                },
                "daily_budget": {
                    "type": "integer"
                },
                "main_action": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.CreateAdCampaignResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "dto.GroupStatsResponse": {
            "type": "object",
            "properties": {
                "ads": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StatsEntityRow"
                    }
                },
                "period": {
                    "$ref": "#/definitions/dto.StatsPeriod"
                },
                "placements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StatsEntityRow"
                    }
                },
                "previous_totals": {
                    "$ref": "#/definitions/dto.StatsMetric"
                },
                "timeline": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StatsPoint"
                    }
                },
                "totals": {
                    "$ref": "#/definitions/dto.StatsMetric"
                },
                "unavailable_metrics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StatsUnavailableMetric"
                    }
                }
            }
        },
        "dto.ListAdCampaignsResponse": {
            "type": "object",
            "properties": {
                "advertiser_id": {
                    "type": "integer"
                },
                "campaigns": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.AdCampaignResponse"
                    }
                }
            }
        },
...
        "dto.StatsEntityRow": {
            "type": "object",
            "properties": {
                "clicks": {
                    "type": "integer"
                },
                "cpc": {
                    "type": "number"
                },
                "ctr": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "impressions": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "partner_reward": {
                    "type": "integer"
                },
                "platform_revenue": {
                    "type": "integer"
                },
                "spend": {
                    "type": "integer"
                }
            }
        },
        "dto.StatsMetric": {
            "type": "object",
            "properties": {
                "clicks": {
                    "type": "integer"
                },
                "cpc": {
                    "type": "number"
                },
                "ctr": {
                    "type": "number"
                },
                "impressions": {
                    "type": "integer"
                },
                "partner_reward": {
                    "type": "integer"
                },
                "platform_revenue": {
                    "type": "integer"
                },
                "spend": {
                    "type": "integer"
                }
            }
        },
        "dto.StatsPeriod": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "dto.StatsPoint": {
            "type": "object",
            "properties": {
                "clicks": {
                    "type": "integer"
                },
                "cpc": {
                    "type": "number"
                },
                "ctr": {
                    "type": "number"
                },
                "date": {
                    "type": "string"
                },
                "impressions": {
                    "type": "integer"
                },
                "partner_reward": {
                    "type": "integer"
                },
                "platform_revenue": {
                    "type": "integer"
                },
                "spend": {
                    "type": "integer"
                }
            }
        },
        "dto.StatsUnavailableMetric": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "httpx.Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    }
}